Реализована компиляция простейших математических операций
Класс Builder:
- Метод Link теперь компонует файлы как консольные приложения
- Добавлена проверка на существование exe-файла перед его перемещением в
целевую папку в методе BuildAll
Класс CodeGenerator:
- Добавлен метод Initialize для пересоздания траслированного кода
- В методе DeclareDataSegment добавлены описания системных функций,
описание размера стека и служебные переменные consoleOutHandle,
bytesWritten, buffer, bufferSize
- Метод DeclareMainProcedureEnding теперь вызывает функцию ExitProcess
из WinAPI вместо команды ret для завершения программы
- Добавлены методы арифметических операций над стековыми значениями
- Добавлена реализация метода Print для вывода не более чем
шестизначного числа на экран
Класс Controller:
- Теперь каждый раз при вызове метода Compile сгенерированный ранее код
будет сбрасываться
Класс Lexems:
- Добавлена лексема Print
Класс LexicalAnalyzer:
- Теперь каждый раз при вызове метода DecodeNextLexem поля класса
currentLexem и currentName будут устанавливаться в своё значение по
умолчанию
- Добавлено ключевое слово Print
Класс NameTable:
- Теперь не будет выводить ошибку, если переданное имя для поиска
переменной имело значение null
Класс Reader:
- При вызове метода ReadNextSymbol чтение нового символа будет
производиться после проверки предыдущего
Класс SyntaxAnalyzer:
- Добавлен метод DecodePrintOperation, распознающий лексему Print после
лексемы End. Наличие Print теперь обязательно для каждой программы
- Уточнены сообщения об ошибках - теперь приводится номер строки, в
которой возникла ошибка
- Добавлены вызовы арифметических методов класса CodeGenerator
Класс MainForm:
- Добавлен пункт меню "Файл - Новый", очищающий содержимое текстовых
полей и путь к текущему файлу
- Текстовые поля resultRichTextBox и compileRichTextBox теперь
предназначены только для чтения

Комментарий: Я научил его считать!

+++ CodeBuilder/Builder.cs
@@ -53,7 +53,7 @@ namespace CodeBuilder
                     Path.GetDirectoryName(fileName), "Compiled",
                     exeFileName);
             if (File.Exists(exeFilePath)) File.Delete(exeFilePath);
-            File.Move(exeFileName, exeFilePath);
+            if (File.Exists(exeFileName)) File.Move(exeFileName, exeFilePath);
         }
 
         /// <summary>
@@ -91,7 +91,7 @@ namespace CodeBuilder
                 StartInfo = new ProcessStartInfo
                 {
                     FileName = "link.exe",
-                    Arguments = "/SUBSYSTEM:WINDOWS /OPT:NOREF " + filename,
+                    Arguments = "/SUBSYSTEM:CONSOLE /OPT:NOREF /defaultlib:kernel32.lib " + filename,
                     UseShellExecute = false,
                     RedirectStandardOutput = true,
                     CreateNoWindow = true
